requirements.txt
streamlit
openai
pypdf
python-docx
python-dotenv
streamlit-webrtc
av
pydub
numpy
